+++ pkgs/ffigen/lib/src/config_provider/config.dart
@@ -180,6 +180,10 @@ final class Input {
   /// Command line arguments to pass to clang_compiler.
   final List<String>? compilerOptions;
 
+  /// Whether [compilerOptions] should be appended to the default compiler
+  /// options, instead of replacing them.
+  final bool appendCompilerOptions;
+
   /// Where to ignore compiler warnings/errors in source header files.
   final bool ignoreSourceErrors;
 
@@ -187,6 +191,7 @@ final class Input {
     this.entryPoints = const [],
     this.include = _includeDefault,
     this.compilerOptions,
+    this.appendCompilerOptions = false,
     this.ignoreSourceErrors = false,
   });
 }