+++ pkgs/ffigen/lib/src/context.dart
@@ -27,9 +27,11 @@ class Context {
   bool hasSourceErrors = false;
   final reportedCommentRanges = <((String, int), (String, int))>{};
   final libs = LibraryImports();
-  late final compilerOpts =
-      config.input.compilerOptions ??
-      defaultCompilerOpts(logger, cpp: config.cpp != null);
+  late final compilerOpts = computeCompilerOpts(
+    input: config.input,
+    logger: logger,
+    cpp: config.cpp != null,
+  );
   final Scope rootScope = Scope.createRoot('root');
   final Scope rootObjCScope = Scope.createRoot('objc_root');
   late final ExtraSymbols extraSymbols;